move test into logic and also add manual resolve test
